+++ src/org/exist/dom/DocumentImpl.java
@@ -563,7 +563,7 @@ public class DocumentImpl extends NodeImpl implements Document, Comparable {
 		final SecurityManager secman = broker.getBrokerPool().getSecurityManager();
 		final int uid = istream.readInt();
 		final int gid = istream.readInt();
-		final int perm = (istream.readByte() & 0777);
+		final int perm = (istream.readInt() & 0777);
 		if (secman == null) {
 			permissions.setOwner(SecurityManager.DBA_USER);
 			permissions.setGroup(SecurityManager.DBA_GROUP);
@@ -686,7 +686,7 @@ public class DocumentImpl extends NodeImpl implements Document, Comparable {
 			ostream.writeInt(user.getUID());
 			ostream.writeInt(group.getId());
 		}
-		ostream.writeByte((byte) permissions.getPermissions());
+		ostream.writeInt(permissions.getPermissions());
 		if(lockOwnerId > 0)
 			ostream.writeInt(lockOwnerId);
 		else
